+++ pkgs/jnigen/test/kotlin_test/runtime_test_registrant.dart
@@ -34,6 +34,13 @@ void registerTests(String groupName, TestRunnerCallback test) {
         expect(noDelayNull, null);
         final asyncNull = await suspendFun.nullableHello(true);
         expect(asyncNull, null);
+
+        expect(suspendFun.getResult(), 0);
+        final voidFuture = suspendFun.noReturn();
+        expect(voidFuture, isA<Future<void>>());
+        expect(voidFuture, isNot(isA<Future<JObject>>()));
+        await voidFuture;
+        expect(suspendFun.getResult(), 123);
       });
     });
 
@@ -358,6 +365,7 @@ void registerTests(String groupName, TestRunnerCallback test) {
 
     group('Interface with suspend functions', () {
       test('return immediately', () async {
+        var result = 0;
         final itf = SuspendInterface.implement($SuspendInterface(
           sayHello: () async => JString.fromString('Hello'),
           sayHello$1: (JString name) async =>
@@ -368,6 +376,7 @@ void registerTests(String groupName, TestRunnerCallback test) {
           sayInt$1: (JInteger value) async => JInteger(10 * value.intValue()),
           nullableInt: (bool returnNull) async =>
               returnNull ? null : JInteger(123),
+          noReturn: () async => result = 123,
         ));
 
         expect((await itf.sayHello()).toDartString(), 'Hello');
@@ -379,6 +388,8 @@ void registerTests(String groupName, TestRunnerCallback test) {
         expect((await itf.sayInt$1(JInteger(456))).intValue(), 4560);
         expect((await itf.nullableInt(false))?.intValue(), 123);
         expect(await itf.nullableInt(true), null);
+        await itf.noReturn();
+        expect(result, 123);
 
         expect(
             (await consumeOnSameThread(itf)).toDartString(),
@@ -389,6 +400,7 @@ Hello
 123
 7890
 123
+kotlin.Unit
 '''
                 .trim());
         expect(
@@ -400,11 +412,13 @@ Hello
 123
 7890
 123
+kotlin.Unit
 '''
                 .trim());
       });
 
       test('return delayed', () async {
+        var result = 0;
         final itf = SuspendInterface.implement($SuspendInterface(
           sayHello: () async {
             await Future<void>.delayed(const Duration(milliseconds: 100));
@@ -430,6 +444,10 @@ Hello
             await Future<void>.delayed(const Duration(milliseconds: 100));
             return returnNull ? null : JInteger(123);
           },
+          noReturn: () async {
+            await Future<void>.delayed(const Duration(milliseconds: 100));
+            result = 123;
+          },
         ));
 
         expect((await itf.sayHello()).toDartString(), 'Hello');
@@ -441,6 +459,8 @@ Hello
         expect((await itf.sayInt$1(JInteger(456))).intValue(), 4560);
         expect((await itf.nullableInt(false))?.intValue(), 123);
         expect(await itf.nullableInt(true), null);
+        await itf.noReturn();
+        expect(result, 123);
 
         expect(
             (await consumeOnSameThread(itf)).toDartString(),
@@ -451,6 +471,7 @@ Hello
 123
 7890
 123
+kotlin.Unit
 '''
                 .trim());
         expect(
@@ -462,6 +483,7 @@ Hello
 123
 7890
 123
+kotlin.Unit
 '''
                 .trim());
       });
@@ -474,6 +496,7 @@ Hello
           sayInt: () async => throw Exception(),
           sayInt$1: (JInteger value) async => throw Exception(),
           nullableInt: (bool returnNull) async => throw Exception(),
+          noReturn: () async => throw Exception(),
         ));
 
         await expectLater(itf.sayHello(), throwsA(isA<JniException>()));
@@ -485,6 +508,7 @@ Hello
         await expectLater(
             itf.sayInt$1(JInteger(456)), throwsA(isA<JniException>()));
         await expectLater(itf.nullableInt(false), throwsA(isA<JniException>()));
+        await expectLater(itf.noReturn(), throwsA(isA<JniException>()));
 
         await expectLater(
             consumeOnSameThread(itf), throwsA(isA<JniException>()));
@@ -518,6 +542,10 @@ Hello
             await Future<void>.delayed(const Duration(milliseconds: 100));
             throw Exception();
           },
+          noReturn: () async {
+            await Future<void>.delayed(const Duration(milliseconds: 100));
+            throw Exception();
+          },
         ));
 
         await expectLater(itf.sayHello(), throwsA(isA<JniException>()));
@@ -529,6 +557,7 @@ Hello
         await expectLater(
             itf.sayInt$1(JInteger(456)), throwsA(isA<JniException>()));
         await expectLater(itf.nullableInt(false), throwsA(isA<JniException>()));
+        await expectLater(itf.noReturn(), throwsA(isA<JniException>()));
 
         await expectLater(
             consumeOnSameThread(itf), throwsA(isA<JniException>()));